refactor: move new helpers around
