+++ packages/api-reference/src/packages/openapi-sdk/utils/createDefaultTag.ts
