+++ packages/api-reference/src/packages/openapi-sdk/utils/filterInternalItems.ts
