+++ packages/api-reference/src/packages/openapi-sdk/utils/getOpenApiDocument.ts
