+++ packages/api-reference/src/packages/openapi-sdk/utils/pending.ts
